third_party/WebKit/LayoutTests/navigator_webdriver/navigator_webdriver.html

<!DOCTYPE html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

// navigator.webdriver property is a runtime enabled feature
// that should not exist if not excplicitly enabled
test(function() {
  assert_false("webdriver" in window.navigator);
  assert_true(navigator.webdriver === undefined);
}, "Test if webdriver property is not in the prototype chain");
</script>
</body>
</html>
